+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+### Added
+
+- Hijri calendars (tabular scheme IIa, Umm al-Qura)
+
 ## <a id="v1.15.4"></a>[v1.15.4] - 2026-09-22
 
 ### Added